@@ -19,6 +19,10 @@
         {
             "name": "Alexandre Caillot (Shiroe_sama)",
             "email": "a.caillot@openium.fr"
+        },
+        {
+            "name": "Ophélie Coudert",
+            "email": "o.coudert@openium.fr"
         }
     ],
     "autoload": {
@@ -30,42 +34,42 @@
         ]
     },
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "ext-json": "*",
         "ext-pdo": "*",
-        "symfony/http-foundation": "^6.0",
+        "symfony/http-foundation": "^7.0",
         "doctrine/orm": "^2.6",
-        "symfony/validator": "^6.0",
-        "symfony/http-kernel": "^6.0",
-        "symfony/dependency-injection": "^6.0",
-        "symfony/config": "^6.0",
-        "symfony/framework-bundle": "^6.0",
-        "symfony/monolog-bridge": "^6.0",
-        "symfony/event-dispatcher": "^6.0",
-        "symfony/translation": "^6.0",
+        "symfony/validator": "^7.0",
+        "symfony/http-kernel": "^7.0",
+        "symfony/dependency-injection": "^7.0",
+        "symfony/config": "^7.0",
+        "symfony/framework-bundle": "^7.0",
+        "symfony/monolog-bridge": "^7.0",
+        "symfony/event-dispatcher": "^7.0",
+        "symfony/translation": "^7.0",
         "symfony/monolog-bundle": "^3.2",
-        "symfony/security-core": "^6.0",
-        "symfony/finder": "^6.0",
-        "symfony/filesystem": "^6.0",
-        "symfony/console": "^6.0",
-        "symfony/security-http": "^6.0",
+        "symfony/security-core": "^7.0",
+        "symfony/finder": "^7.0",
+        "symfony/filesystem": "^7.0",
+        "symfony/console": "^7.0",
+        "symfony/security-http": "^7.0",
         "symfony/orm-pack": "^2.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.0",
-        "symfony/phpunit-bridge": "^6.0",
-        "symfony/cache": "^6.0",
-        "phpstan/phpstan": "^1.9",
-        "symfony/error-handler": "^6.0",
+        "phpunit/phpunit": "^10.0",
+        "symfony/phpunit-bridge": "^7.0",
+        "symfony/cache": "^7.0",
+        "phpstan/phpstan": "^1.10",
+        "symfony/error-handler": "^7.0",
         "phpstan/extension-installer": "^1.2",
         "phpstan/phpstan-phpunit": "^1.2",
         "phpstan/phpstan-strict-rules": "^1.4",
-        "rector/rector": "^0.16.0"
+        "rector/rector": "^0.18.13"
     },
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "^6.0"
+            "require": "^7.0"
         },
         "phpstan": {
             "includes": [
