@@ -28,13 +28,13 @@ return static function (RectorConfig $rectorConfig): void {
     // define sets of rules
     $rectorConfig->sets(
         [
-            LevelSetList::UP_TO_PHP_81,
+            LevelSetList::UP_TO_PHP_82,
             SetList::DEAD_CODE,
             SetList::CODE_QUALITY,
-            SymfonySetList::SYMFONY_60,
+            SymfonySetList::SYMFONY_64,
             SymfonySetList::SYMFONY_CODE_QUALITY,
             SymfonySetList::SYMFONY_CONSTRUCTOR_INJECTION,
-            PHPUnitSetList::PHPUNIT_90,
+            PHPUnitSetList::PHPUNIT_100,
         ]
     );
 };
